+++ packages/fluo-vector/test/fluoVector.color.test.js
@@ -1,24 +1,11 @@
-import { RENDER }           from '@palett/enum-colorant-modes'
-import { BOLD, ITALIC }     from '@palett/enum-font-effects'
-import { PresetCollection } from '@palett/fluo'
-import { FRESH }            from '@palett/presets'
-import { logger }           from '@spare/logger'
-import { fluoVector }       from '../index.js'
-import { candidates }       from './strategies/candidates'
+import { FRESH }      from '@palett/presets'
+import { test }       from 'node:test'
+import { fluoVector } from '../index.js'
+import { candidates } from './strategies/candidates.js'
 
-for (let [k, vector] of Object.entries(candidates)) {
-  fluoVector.call(
-    {
-      mutate: false,
-      colorant: RENDER
-    },
-    vector,
-    PresetCollection.build(FRESH).assignEffect(BOLD, ITALIC)
-    // [
-    //   { preset: FRESH, by: isNumeric, to: x => x },
-    //   { preset: PLANET, by: isLiteral, to: stringValue }
-    // ]
-  )
-    |> logger
-  // |> says[k]
-}
+test('fluoVector colors', () => {
+  for (let [ k, vector ] of Object.entries(candidates)) {
+    console.log(k, fluoVector(vector, FRESH).join(', '))
+  }
+})
+