+++ constants/enum-colorant-modes/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @palett/enum-colorant-modes
 
+## 0.12.1
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @palett/enum-color-space@0.12.1
+
 ## 0.12.0
 
 ### Minor Changes