+++ constants/enum-font-effects/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @palett/enum-font-effects
 
+## 0.12.1
+
+### Patch Changes
+
+- minor fix
+
 ## 0.12.0
 
 ### Minor Changes