+++ packages/fluo-matrix/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @palett/fluo-matrix
 
+## 0.12.1
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @palett/fluo-vector@0.12.1
+  - @palett/convert@0.12.1
+  - @palett/fluo@0.12.1
+  - @palett/enum-colorant-modes@0.12.1
+  - @palett/projector@0.12.1
+
 ## 0.12.0
 
 ### Minor Changes