+++ packages/munsell/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @palett/munsell
 
+## 0.12.1
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @palett/color-space@0.12.1
+  - @palett/convert@0.12.1
+
 ## 0.12.0
 
 ### Minor Changes