+++ packages/projector/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @palett/projector
 
+## 0.12.1
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @palett/dye@0.12.1
+  - @palett/presets@0.12.1
+
 ## 0.12.0
 
 ### Minor Changes