+++ packages/stringify/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @palett/stringify
 
+## 0.12.1
+
+### Patch Changes
+
+- minor fix
+- Updated dependencies
+  - @palett/convert@0.12.1
+  - @palett/enum-ansi-codes@0.12.1
+  - @palett/util-ansi@0.12.1
+
 ## 0.12.0
 
 ### Minor Changes